Rewrite for gbit compilation
