+++ .gitignore
@@ -1,4 +1,3 @@
 build/
 *.o
 *.o.d
-*.so
