+++ src/emit.dasc
@@ -1,7 +1,7 @@
 #include "../src/emit.h"
 
-#ifdef GBIT
-#include "../gbit_src/test_cpu.h"
+#ifdef INSTRUCTION_TEST
+#include "tests/instr_test.h"
 #endif
 
 /* Configure the generator */
@@ -128,7 +128,7 @@ static bool inst_jp(dasm_State **Dst, gbz80_inst *inst, uint64_t *cycles)
         | dec SP
         | and xSP, 0xffff
         | mov word [aMem + xSP], (uint16_t)(inst->address + inst->bytes)
-#ifdef GBIT
+#ifdef INSTRUCTION_TEST
         | mov tmp1, xSP
         | ld16 tmp1, (inst->address + inst->bytes)
 #endif
@@ -349,7 +349,7 @@ static bool inst_ld16(dasm_State **Dst, gbz80_inst *inst, uint64_t *cycles)
             uint16_t addr = (uint16_t)(inst->args[1] + 256 * inst->args[2]);
             | mov word [aMem + addr], SP;
             | mov tmp1, xSP
-#ifdef GBIT
+#ifdef INSTRUCTION_TEST
             | ld16 addr, tmp1
 #endif
         } else {
@@ -391,7 +391,7 @@ static bool inst_push(dasm_State **Dst, gbz80_inst *inst, uint64_t *cycles) {
         | dec SP
         | mov [aMem + xSP + 1], A
         | mov [aMem + xSP], tmp3b
-#ifdef GBIT
+#ifdef INSTRUCTION_TEST
         | mov tmp1, xSP
         | write_byte tmp1, tmp3
         | inc tmp1
@@ -403,7 +403,7 @@ static bool inst_push(dasm_State **Dst, gbz80_inst *inst, uint64_t *cycles) {
         | dec SP
         | mov [aMem + xSP + 1], B
         | mov [aMem + xSP], C
-#ifdef GBIT
+#ifdef INSTRUCTION_TEST
         | mov tmp1, xSP
         | write_byte tmp1, xC
         | inc tmp1
@@ -415,7 +415,7 @@ static bool inst_push(dasm_State **Dst, gbz80_inst *inst, uint64_t *cycles) {
         | dec SP
         | mov [aMem + xSP + 1], D
         | mov [aMem + xSP], E
-#ifdef GBIT
+#ifdef INSTRUCTION_TEST
         | mov tmp1, xSP
         | write_byte tmp1, xE
         | inc tmp1
@@ -427,7 +427,7 @@ static bool inst_push(dasm_State **Dst, gbz80_inst *inst, uint64_t *cycles) {
         | dec SP
         | mov [aMem + xSP + 1], H
         | mov [aMem + xSP], L
-#ifdef GBIT
+#ifdef INSTRUCTION_TEST
         | mov tmp1, xSP
         | mov tmp2 , xL
         | write_byte tmp1, tmp2
@@ -1346,7 +1346,7 @@ static bool inst_swap(dasm_State **Dst, gbz80_inst *inst, uint64_t *cycles)
         | shl byte [aMem + tmp1], 4
         | shr tmp2b, 4
         | or [aMem + tmp1], tmp2b
-#ifdef GBIT
+#ifdef INSTRUCTION_TEST
         | write_byte, tmp1, [aMem + tmp1]
 #endif
         break;
@@ -1393,7 +1393,7 @@ static bool inst_stop(dasm_State **Dst, gbz80_inst *inst, uint64_t *cycles)
     return true;
 }
 
-#ifdef GBIT
+#ifdef INSTRUCTION_TEST
 static bool inst_set_flag(dasm_State **Dst, gbz80_inst *inst, uint64_t *cycles)
 {
     | mov ah, inst->args[0] 
@@ -1442,7 +1442,7 @@ bool emit(gb_block *block, GList *inst)
         }
 
         switch (DATA(inst)->opcode) {
-#ifdef GBIT        
+#ifdef INSTRUCTION_TEST        
         case LD_F:
             if (!inst_load_flag(Dst, DATA(inst), &cycles))
                 goto exit_fail;
